refactor: move extractKeyInfo to its own file
